file structure refactor. placed files to their designated folders and created files for the variants.
