+++ my-app/assets/workData.tsx
@@ -27,20 +27,32 @@ export const WORK__DATA = [
     launch: "https://the-toad-tribune-659c7.web.app/",
   },
 ];
+
+// export const WORK_BACKGROUND = [
+//   {
+//     pokePalace: "bg-gradient-to-r from-[#ffb7005d] to-[#eb412ad2]"
+//   },
+//   {
+//     movieApp: "bg-gradient-to-r from-[#aa03ac2f] to-[#9305ffeb]"
+//   },
+//   {
+//     toadTribune: "bg-gradient-to-r from-[#0ddd8643] to-[#2ecbc0]"
+//   },
+// ]
 export const OTHER__WORK__DATA = [
   {
-    title: "Kovee",
+    title: "kovee",
     description:
       "Lorem, ipsum dolor sit amet consectetur adipisicing elit. Quasi nisi esse adipisci dolores, eligendi voluptates voluptatibus nulla cumque sequi dolore!",
     background: "bg-gradient-to-r from-[#ffb7005d] to-[#eb412ad2]",
-    stacks: ["native", "rtk", "styled components"],
+    stacks: ["native", "rtk query", "styled components"],
     github: "https://github.com/BlackCubes/covidmap-react-native",
   },
   {
     title: "portfolio",
     description:
       "Lorem, ipsum dolor sit amet consectetur adipisicing elit. Quasi nisi esse adipisci dolores, eligendi voluptates voluptatibus nulla cumque sequi dolore!",
-    background: "bg-gradient-to-r from-[#830edd55] to-[#1041e1eb]",
+    background: "bg-gradient-to-r from-[#eb412ad2] to-[#ffb7005d]",
     stacks: ["react", "redux", "sass"],
     github: "https://github.com/JayDampitan/portfolio",
     launch: "https://jaydampitan.com/",
