+++ my-app/components/about/About.tsx
@@ -1,25 +1,13 @@
-import {  useRef } from "react";
+import { useRef } from "react";
 import { motion } from "framer-motion";
 import Link from "next/link";
-import TitleSection from "../components/TitleSection";
-import { ABOUT__DATA, STACKS__DATA, SERVICES__DATA } from "../assets/data";
+import TitleSection from "../utils/TitleSection";
+import { ABOUT__DATA, STACKS__DATA, SERVICES__DATA } from "../../assets/data";
 import { AiFillCheckSquare } from "react-icons/ai";
+import { whiteVariant } from "./aboutVariant"
 
 type Props = {};
-const whiteVariant = {
-  visible: {
-    y: [0, 10, 10, 30, 30, 10, 10, 0],
-    scale: [1, 1.2, 1, 1.2, 1, 1.2, 1],
-    transition: {
-      duration: 8,
-      delay: 5,
-      repeat: Infinity,
-    },
-  },
-  hover: {
-    scale: 1.2,
-  },
-};
+
 
 const About = (props: Props) => {
   const constraintRef = useRef(null);
@@ -29,7 +17,7 @@ const About = (props: Props) => {
       className=" h-screen w-full bg-darkGrey flex items-center justify-center relative px-4"
     >
       <div className=" h-screen max-w-[1700px] justify-center px-4 w-full flex items-center flex-col border-b-[1px] border-white ">
-        <TitleSection  title="something about me." />
+        <TitleSection title="something about me" />
         <div className=" flex flex-col items-center justify-center ">
           <div className=" py-2 xl:w-[50%]  lg:max-w-[1700px]">
             {ABOUT__DATA.map((item, index) => (
@@ -42,7 +30,7 @@ const About = (props: Props) => {
             ))}
           </div>
           {/* ---------- skills -------------- */}
-          <div className="flex-wrap flex gap-3 lg:gap-6 pt-6 max-w-[500px] lg:max-w-[600px]">
+          <div className="flex-wrap flex gap-3 lg:gap-6 pt-10 max-w-[500px] lg:max-w-[600px]">
             {STACKS__DATA.map((item, index) => (
               <ul key={index}>
                 <li className="text-white flex items-center">
@@ -58,14 +46,19 @@ const About = (props: Props) => {
         </div>
 
         {/* ---------- services offered -------------- */}
-        <div className=" flex gap-4  mt-14 lg:pr-20">
+        <div className=" flex flex-col lg:flex-row gap-4  mt-14 w-2/3">
           {SERVICES__DATA.map((item, index) => (
+            <>
             <div
               key={index}
-              className="text-darkGrey text-sm lg:text-lg bg-white rounded-md  p-1 lg:px-4 lg:py-3 border border-white text-center flex items-center hover:-translate-y-1 duration-200"
+              className=" text-sm text-white rounded-md  p-1 lg:px-4 lg:py-3 border border-white text-center flex flex-col items-center hover:-translate-y-1 duration-200"
             >
-              {item}
+             <h4 className="text-lg font-semibold"><span className="px-2">hello</span>{item.skill}</h4>
+            <p className="leading-7">
+              {item.description}
+            </p>
             </div>
+            </>
           ))}
         </div>
       </div>
