+++ my-app/components/about/aboutVariant.tsx
@@ -1,0 +1,14 @@
+export const whiteVariant = {
+    visible: {
+      y: [0, 10, 10, 30, 30, 10, 10, 0],
+      scale: [1, 1.2, 1, 1.2, 1, 1.2, 1],
+      transition: {
+        duration: 8,
+        delay: 5,
+        repeat: Infinity,
+      },
+    },
+    hover: {
+      scale: 1.2,
+    },
+  };
