+++ my-app/components/contact/Contact.tsx
@@ -1,71 +1,31 @@
 import React, { useRef } from "react";
-import TitleSection from "./TitleSection";
-import ToTop from "./ToTop";
+import TitleSection from "../utils/TitleSection";
+import ToTop from "../utils/ToTop";
 import { motion } from "framer-motion";
 import Link from "next/link";
 import { BiPlanet } from "react-icons/bi";
-import SmallCirlces from "./SmallCircles";
+import SmallCirlces from "../utils/SmallCircles";
+import {whiteVariant, orangeVariant, blueVariant } from "./contactVariant"
 
 type Props = {};
-
-const whiteVariant = {
-  visible: {
-    y: [0, 10, 10, 30, 30, 10, 10, 0],
-    scale: [1, 1.1, 1, 1.1, 1, 1.1, 1],
-    transition: {
-      duration: 8,
-      delay: 5,
-      repeat: Infinity,
-    },
-  },
-  hover: {
-    scale: 1.2,
-  },
-};
-const blueVariant = {
-  visible: {
-    y: [0, 10, 10, 30, 30, 10, 10, 0],
-    scale: [1, 1.1, 1, 1.1, 1, 1.1, 1],
-    transition: {
-      duration: 8,
-      delay: 5.5,
-      repeat: Infinity,
-    },
-  },
-  hover: {
-    scale: 1.2,
-  },
-};
-const yellowVariant = {
-  visible: {
-    y: [0, 10, 10, 30, 30, 10, 10, 0],
-    scale: [1, 1.1, 1, 1.1, 1, 1.1, 1],
-    transition: {
-      duration: 8,
-      delay: 6,
-      repeat: Infinity,
-    },
-  },
-  hover: {
-    scale: 1.2,
-  },
-};
 
 const Contact = (props: Props) => {
   const constraintsRef = useRef(null);
   return (
     <div className="h-screen flex flex-col items-center justify-center bg-darkGrey relative">
-      <TitleSection  title="say hello." />
+      <TitleSection  title="say hello" />
       <div ref={constraintsRef} className="mt-20  relative">
         <motion.div
           animate={{ scale: [1, 1.1, 1, 1.1, 1, 1.1, 1] }}
           transition={{ duration: 8, repeat: Infinity }}
-          className="w-[150px] h-[150px] lg:w-[300px] lg:h-[300px] flex items-center justify-center rounded-full bg-[#C5A639] "
+          className="w-[150px] h-[150px] lg:w-[300px] lg:h-[300px] flex items-center justify-center rounded-full bg-yellow "
         >
           <h2 className="text-xl lg:text-4xl text-gray-500 hover:text-white hover:scale-105 cursor-pointer transition ease-in duration-400">
             jaydampitan@gmail.com
           </h2>
         </motion.div>
+
+        {/* ---------background blurs----------- */}
         <motion.div
           variants={whiteVariant}
           whileHover="hover"
@@ -75,12 +35,12 @@ const Contact = (props: Props) => {
           className="absolute w-[40%] h-[40%] bg-white -top-4 -right-9  blur-3xl rounded-full cursor-pointer"
         ></motion.div>
         <motion.div
-          variants={yellowVariant}
+          variants={orangeVariant}
           whileHover="hover"
           animate="visible"
           drag
           dragConstraints={constraintsRef}
-          className="absolute w-[40%] h-[40%] bg-yellow-100 -top-4 -left-9  blur-3xl rounded-full cursor-pointer"
+          className="absolute w-[40%] h-[40%] bg-orange-100 -top-4 -left-9  blur-3xl rounded-full cursor-pointer"
         ></motion.div>
         <motion.div
           variants={blueVariant}
@@ -92,6 +52,8 @@ const Contact = (props: Props) => {
         ></motion.div>
       </div>
       <ToTop />
+
+      {/* -------------planet feature------------ */}
       <div className="absolute left-3 bottom-3 lg:bottom-9 lg:left-9 flex gap-4 hover:scale-110 ">
         <div className="w-[20px] h-[20px] flex items-center justify-center border rounded-full cursor-pointer">
           <Link
@@ -104,6 +66,7 @@ const Contact = (props: Props) => {
           <SmallCirlces/>
         </div>
       </div>
+      {/* -----------end of planet feature------------ */}
     </div>
   );
 };
