+++ my-app/components/contact/contactVariant.tsx
@@ -1,0 +1,44 @@
+export const whiteVariant = {
+    visible: {
+      y: [0, 10, 10, 30, 30, 10, 10, 0],
+      scale: [1, 1.1, 1, 1.1, 1, 1.1, 1],
+      transition: {
+        duration: 8,
+        delay: 5,
+        repeat: Infinity,
+      },
+    },
+    hover: {
+      scale: 1.2,
+    },
+  };
+
+  export const blueVariant = {
+    visible: {
+      y: [0, 10, 10, 30, 30, 10, 10, 0],
+      scale: [1, 1.1, 1, 1.1, 1, 1.1, 1],
+      transition: {
+        duration: 8,
+        delay: 5.5,
+        repeat: Infinity,
+      },
+    },
+    hover: {
+      scale: 1.2,
+    },
+  };
+
+  export const orangeVariant = {
+    visible: {
+      y: [0, 10, 10, 30, 30, 10, 10, 0],
+      scale: [1, 1.1, 1, 1.1, 1, 1.1, 1],
+      transition: {
+        duration: 8,
+        delay: 6,
+        repeat: Infinity,
+      },
+    },
+    hover: {
+      scale: 1.2,
+    },
+  };
