+++ my-app/components/nav/navVariants.tsx
@@ -1,0 +1,12 @@
+export const containerVariant = {
+  hidden: {
+    opacity: 0,
+  },
+  visible: {
+    opacity: 1,
+    transition: {
+      duration: 2,
+      delay: 5,
+    },
+  },
+};
