+++ my-app/components/utils/scrollDown/scrollDownVariants.tsx
@@ -1,0 +1,14 @@
+export const containerVariant = {
+    hidden: {
+      opacity: 0,
+      y: -50,
+    },
+    visible: {
+      opacity: 1,
+      y: 0,
+      transition: {
+        duration: 2,
+        delay: 5,
+      },
+    },
+  };
